Moved lazy.nvim to Vim packages
