--- lua/config/init.lua
@@ -1,3 +1,0 @@
-require("config.options")
-require("config.keymaps")
-require("config.lazy")
