--- lua/plugins/abyss.lua
@@ -1,7 +1,0 @@
-return {
-    "barrientosvctor/abyss.nvim",
-    dev = true,
-    lazy = false,
-    priority = 1000,
-    opts = {}
-}
