--- lua/plugins/autopairs.lua
@@ -1,5 +1,0 @@
-return {
-    "windwp/nvim-autopairs",
-    event = "InsertEnter",
-    config = true
-}
