--- lua/plugins/comment.lua
@@ -1,17 +1,0 @@
-return {
-    "numToStr/Comment.nvim",
-    lazy = true,
-    opts = {},
-    keys = {
-        -- Visual mode
-        { "gc", mode = "v" },
-        { "gb", mode = "v" },
-
-        -- Normal mode
-        { "gcc", mode = "n" },
-        { "gbc", mode = "n" },
-        { "gcO", mode = "n" },
-        { "gco", mode = "n" },
-        { "gcA", mode = "n" },
-    },
-}
