+++ plugin/autopairs.lua
@@ -1,0 +1,1 @@
+require("nvim-autopairs").setup {}
