+++ plugin/cmp.lua
@@ -1,0 +1,80 @@
+local cmp = require 'cmp'
+local cmp_autopairs = require('nvim-autopairs.completion.cmp')
+local luasnip = require 'luasnip'
+
+luasnip.config.set_config {
+    history = true,
+    enable_autosnippets = true,
+}
+
+-- NOTE: the loader is called twice so it picks up the defaults first then my custom vscode snippets.
+-- see: https://github.com/L3MON4D3/LuaSnip/issues/364
+require('luasnip.loaders.from_vscode').lazy_load()
+require('luasnip.loaders.from_vscode').lazy_load({ paths = "./snippets" })
+
+cmp.setup {
+    formatting = {
+        format = function(entry, vim_item)
+            vim_item.kind = string.format('%s', vim_item.kind) -- This concatenates the name of the item kind
+            -- Source
+            vim_item.menu = ({
+                buffer = "[Buffer]",
+                nvim_lsp = "[LSP]",
+                luasnip = "[LuaSnip]",
+                nvim_lua = "[Lua]",
+                latex_symbols = "[LaTeX]",
+            })[entry.source.name]
+            return vim_item
+        end
+    },
+    snippet = {
+        expand = function(args)
+            luasnip.lsp_expand(args.body)
+        end,
+    },
+    completion = {
+        completeopt = 'menu,menuone,noinsert'
+    },
+    mapping = cmp.mapping.preset.insert {
+        ['<C-z>'] = cmp.mapping.complete {},
+        ['<CR>'] = cmp.mapping(function(fallback)
+            if cmp.visible() then
+                if luasnip.expandable() then
+                    luasnip.expand()
+                else
+                    cmp.confirm({
+                        behavior = cmp.ConfirmBehavior.Replace,
+                        select = true
+                    })
+                end
+            else
+                fallback()
+            end
+        end),
+        ["<Tab>"] = cmp.mapping(function(fallback)
+            if cmp.visible() then
+                cmp.select_next_item()
+            elseif luasnip.locally_jumpable(1) then
+                luasnip.jump(1)
+            else
+                fallback()
+            end
+        end, { "i", "s" }),
+        ['<S-Tab>'] = cmp.mapping(function(fallback)
+            if cmp.visible() then
+                cmp.select_prev_item()
+            elseif luasnip.locally_jumpable(-1) then
+                luasnip.jump(-1)
+            else
+                fallback()
+            end
+        end, { 'i', 's' }),
+    },
+    sources = {
+        { name = 'nvim_lsp' },
+        { name = 'luasnip' },
+    },
+}
+
+-- Add parentheses after selecting function or method item
+cmp.event:on('confirm_done', cmp_autopairs.on_confirm_done())
