+++ plugin/comment.lua
@@ -1,0 +1,1 @@
+require('Comment').setup()
